% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean Functions.R
\name{clean_lineup_data}
\alias{clean_lineup_data}
\title{Clean DraftKings 'contest-standings' csv Lineup Data}
\usage{
clean_lineup_data(contest_data, sport)
}
\arguments{
\item{contest_data}{A dataframe of raw lineup data from a draftings "contest-standings" csv}

\item{sport}{A string specifying the sport of the contest. tidyDK currently supports
'NFL', 'NHL', 'NBA' and 'MLB'}
}
\value{
The contest data for all lineups in the 'contest-standings' csv in long format.
 Returns a data.table object
}
\description{
Takes a draftkings "contest-standings" dataframe with player lineups and returns
a tidy version in a "long" format i.e. each row is 1 player corresponding to a lineup id.
}
\examples{
\donttest{

lineups <- clean_lineup_data(contest_data, "NFL")
head(lineups)

}

}
