% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean Functions.R
\name{clean_player_contest_info}
\alias{clean_player_contest_info}
\title{Clean DraftKings player information (ownership, points scored) from a
DraftKings'contest-standings' csv}
\usage{
clean_player_contest_info(player_contest_data)
}
\arguments{
\item{player_contest_data}{A dataframe of player ownership and points scored information from a
DraftKings "contest-standings" csv (columns 8-11)}
}
\value{
Player information for a given contest with numeric variables
}
\description{
Takes a dataframe with player ownership and points information from
a draftkings 'contest-standings' csv, and returns a clean dataframe with values
converted to numeric
}
\examples{
\donttest{

player_info <- clean_player_contest_info(player_data)
head(player_info)

}
}
