% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean Functions.R
\name{clean_salary}
\alias{clean_salary}
\title{Clean DraftKings salary data}
\usage{
clean_salary(salary_dat, sport)
}
\arguments{
\item{player_data}{A dataframe of player ownership and points scored information from a
DraftKings "contest-standings" csv (columns 8-11)}
}
\value{
Player information for a given contest with numeric variables
}
\description{
Takes a dataframe with salary data from a DraftKings'DKSalaries' csv
and returns a clean dataframe
}
\examples{
\donttest{

salary_clean <- clean_player_contest_info(salary_data)
head(salary_clean)

}
}
