% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Ownership For Subset of DFS Players.R
\name{get_dk_user_subset_ownership}
\alias{get_dk_user_subset_ownership}
\title{Get Ownership for lineups from a subset of DK users}
\usage{
get_dk_user_subset_ownership(tidy_lus, dk_user_subset)
}
\arguments{
\item{tidy_lus}{long format, tidy lineup data}

\item{dk_user_subset}{a vector of DK usernames - note: NOT CASE SENSITIVE}
}
\value{
a dataframe with ownership for all players in the contest, with total ownership in contest,
 and ownership for the subset of dk usernames provided to the function.
}
\description{
Calculates player ownership for subset of lineups from DK users.
 Can be used to compare sharp ownership vs contest.
}
\examples{
\donttest{

 own_comparison <- get_dk_user_subset_ownership(lu_dat, c('awesemo', 'youdacao', 'papagates', 'rynpak')) 

 own_comparison 

}

}
