% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Lineup Summary.R
\name{get_lineup_summary}
\alias{get_lineup_summary}
\title{Get Lineup Summary}
\usage{
get_lineup_summary(
  tidy_lu,
  analysis_type = "live",
  stack_summary = F,
  sport = NULL
)
}
\arguments{
\item{tidy_lu}{a data.frame of tidy lineup data obtained from the get_tidy_lineups() function}

\item{analysis_type}{"live" for contests that have games which have not started. "post" for contests that have finished.}

\item{stack_summary}{Boolean value to indicate if a stack summary should be included in data 
(NOTE: Currently only NFL is supported, and this will not work for "live" analysis)}

\item{sport}{a character string with one of the sports supported by this package
(NFL, NBA, MLB, NHL)}
}
\value{
tidy data.table with a row for each player in a lineup in a contest,
 joined with ownership and points scored data for each player
}
\description{
Takes tidy long data from get_tidy_lineups() function and creates lineup
  summary of total salary and ownership and other statisitics.  
  
  For contests that are on-going, the function will calculate the remaining salary for LOCKED players.
}
\examples{
\donttest{

lineups <- get_tidy_lineups('./contest-standings-1234564789.csv',
                            './DKsalaries.csv', "NFL")
                            
lu_summary <- get_lineup_summary(tidy_lu, analysis_type = 'post', stack_summary = T, sport = "NFL")                             

head(lu_summary)

}

}
