% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Salary from DB.R
\name{get_salary_db}
\alias{get_salary_db}
\title{Get Salary data from DB}
\usage{
get_salary_db(file_contest_id, sport)
}
\arguments{
\item{file_contest_id}{unique contest identifier used to identify draft group}
}
\description{
Retrieves Salary information from DK API stored in local db
}
