% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Tidy Lineups.R
\name{get_tidy_lineups}
\alias{get_tidy_lineups}
\title{Get Tidy Lineups}
\usage{
get_tidy_lineups(contest_path, salary_path, sport, contest_id = NA)
}
\arguments{
\item{contest_path}{a path to the contest-standings csv you'd like to tidy up}

\item{salary_path}{a path to the salary file corresponding to the contest-standing file you are tidying}

\item{sport}{A string specifying the sport of the contest. tidyDK currently supports
'NFL', 'NHL', 'NBA' and 'MLB'}

\item{contest_id}{an ID that you'd like to assign to all rows of the file (optional)}
}
\value{
tidy data.table with a row for each player in a lineup in a contest,
 joined with ownership and points scored data for each player
}
\description{
Takes paths for draftkings contest-standings file, and corresponding salary file.
  returns a tidy version, where each row corresponds to a player in a lineup.

  Note that this will result in a dataframe of length (number of players in a lineup) * (number of lineups), so
  for larger contests this can result in data sets with 1 million+ rows
}
\examples{
\donttest{

lineups <- get_tidy_lineups('./contest-standings-1234564789.csv',
                            './DKsalaries.csv', "NFL")
head(lineups)

}

}
