% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Tidy Lineups DK API data.R
\name{get_tidy_lineups_dk_api}
\alias{get_tidy_lineups_dk_api}
\title{Get Tidy Lineups DK API}
\usage{
get_tidy_lineups_dk_api(contest_path, sport)
}
\arguments{
\item{contest_path}{a path to the contest-standings csv you'd like to tidy up}

\item{sport}{A string specifying the sport of the contest. tidyDK currently supports
'NFL', 'NHL', 'NBA' and 'MLB'}
}
\value{
tidy data.table with a row for each player in a lineup in a contest,
 joined with ownership and points scored data for each player
}
\description{
Takes paths for draftkings contest-standings file, and gathers salary data 
  from database to create long format lineup data.

  Note that this will result in a dataframe of length (number of players in a lineup) * (number of lineups), so
  for larger contests this can result in data sets with 1 million+ rows
}
\examples{
\donttest{
lineups <- get_tidy_lineups('./contest-standings-1234564789.csv')
head(lineups)
}

}
